+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Added
+- added Content-type-mismatch-to-reponse-body.php for testing new passive rule 1204704.
 
 ## [0.0.2] - 2023-05-30